feat: update changelog for version 1.3.2; fix clear command state management issue; refactor terminal command limits
