+++ src/lib/components/index.ts
@@ -12,7 +12,7 @@ export {default as Help} from './help.svelte';
 export {default as Ls} from './home/terminal-output/ls.svelte';
 
 export {default as ErrorCodes} from './terminal/ErrorCodes.svelte';
-export {default as Modal} from './modals/modal.svelte';
+export {default as Modal} from './modals/Modal.svelte';
 export {default as ModalTitle} from './modals/ModalTitle.svelte';
 
 export {default as List} from './home/terminal-output/Accordion.svelte';
