+++ src/lib/components/terminal/Terminal.svelte
@@ -18,7 +18,7 @@
 			if (text !== null) text.textContent = "";
 			clear.set(false);
 		}
-		// limit on past_commands (clear if > LIMIT_PAST)
+		// limit on past_commands
 		if ($past_commands.length > LIMIT_PAST) {
 			let date = new Date().toISOString().replace('T', ' ').slice(0, 19);
 			past_commands.set([[$username, $pwd, '', empty, true, date]]);
